refactor: DDD 적용을 위한 폴더 구조 리팩토링
